+++ Frontend/src/App.jsx
@@ -2,12 +2,12 @@ import "./App.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import RegisterN from "./pages/Register/Ngo/Registern";
-import RegisterI from "./pages/Register/INDIVISUAL/RegisterI";
-import RegisterS from "./pages/Register/Shop/RegisterS";
-import LoginN from "./pages/Login/NGO/LoginN";
-import LoginI from "./pages/Login/USER/LoginI";
-import LoginS from "./pages/Login/SHOP/LoginS";
+import RegisterN from "./pages/Register/NGO/ngo_register";
+import RegisterI from "./pages/Register/INDIVIDUAL/Individual_Register";
+import RegisterS from "./pages/Register/Shop/shop_register";
+import LoginN from "./pages/Login/NGO/ngo_login";
+import LoginI from "./pages/Login/USER/user_login";
+import LoginS from "./pages/Login/SHOP/shop_login";
 import Allevents from "./pages/ngoevents/Allevents";
 import Campaigns from "./pages/Campaigns/Campaigns";
 import Home from "./pages/Home/Home";
@@ -22,7 +22,7 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Home />} />
-          <Route path="/ngo_register" element={<Registern />} />
+          <Route path="/ngo_register" element={<RegisterN />} />
           <Route path="/ind_register" element={<RegisterI />} />
           <Route path="/shop_register" element={<RegisterS />} />
           <Route path="/ngo_login" element={<LoginN />} />
